+++ .phan/config.php
@@ -197,6 +197,7 @@ return [
     // here to inhibit them from being reported
     'suppress_issue_types' => [
         'PhanUnreferencedClosure',  // False positives seen with closures in arrays, TODO: move closure checks closer to what is done by unused variable plugin
+        'PhanUnreferencedPublicMethod',  // False positives seen in traits
         // 'PhanUndeclaredMethod',
     ],
 
@@ -468,6 +469,8 @@ return [
         'DuplicateArrayKeyPlugin',
         'PregRegexCheckerPlugin',
         'PrintfCheckerPlugin',
+
+        // 'PHPUnitNotDeadCodePlugin',  // Marks phpunit test case subclasses and test cases as refernced code. only useful for runs when dead code detection is enabled
 
         // NOTE: This plugin only produces correct results when
         //       Phan is run on a single core (-j1).