+++ src/Phan/CLI.php
@@ -115,6 +115,7 @@ class CLI
                 'output:',
                 'output-mode:',
                 'parent-constructor-required:',
+                'plugin:',
                 'print-memory-usage-summary',
                 'processes:',
                 'progress-bar',
@@ -328,6 +329,12 @@ class CLI
                 case 'disable-plugins':
                     // Slightly faster, e.g. for daemon mode with lowest latency (along with --quick).
                     Config::setValue('plugins', []);
+                    break;
+                case 'plugin':
+                    Config::setValue(
+                        'plugins',
+                        array_unique(array_merge(Config::getValue('plugins'), [$value]))
+                    );
                     break;
                 case 'use-fallback-parser':
                     Config::setValue('use_fallback_parser', true);
@@ -678,6 +685,10 @@ Usage: {$argv[0]} [options] [files...]
 
  --disable-plugins
   Don't run any plugins. Slightly faster.
+
+ --plugin <pluginName|path/to/Plugin.php>
+  Add an additional plugin to run. This flag can be repeated.
+  (Either pass the name of the plugin or a relative/absolute path to the plugin)
 
  --use-fallback-parser
   If a file to be analyzed is syntactically invalid